去除小数点，改成long类型

+++ app/src/main/java/com/example/brimo/bean/LogBean.java
@@ -7,6 +7,6 @@ public class LogBean {
     private long id;
     private String transaksi;//BFST708101012001508ANDHIKA RYAN:BNINIDJA
     private String time;//18:41:46 WIB
-    private double money;//+ Rp50.000,00
+    private long money;//+ Rp50.000,00
     private String md5;//上面信息进行md5
 }
